Add k8s-job deletion to deploy command
As described in #2 and related to NIVANorge/nivacloud#131 we should
delete jobs before reapplying their .yaml files. Since there already
was a complicated regexp to match the names of actual deployments that
seemed like a good principle to extend.

I splitted out the part of the regexp that actually matches the name.
It does this by recognizing the following structure:
metadata:
  name: Anamewitha-possiblestripe
It only return the nameing part by using the \K keyword in the regexp.
So the implementation is as follows: For each container we are supplied
with the relevant directory that contains it's deployment or job.
For each file in that directory we try to find the
kind: Job
pattern. If we find it we match the name and call
kubectl delete jobs/name

After this we substitute the image name and call
kubectl apply -f - {subbed file contents}

Then if the newly applied file was an actual deployment we add it's
name to a file. After a files have been applied to the k8s cluster we
will use
kubectl rollout status deployment {name}
to wait for the deployments to complete rolling out